sensor: add basic daemon scaffolding
This can start, stop, restart and get status for a simple daemon
process. The daemon sets up some logging for itself and enters a simple
loop for now.
Customer handlers for SIGTERM (just log the signal and exit) and SIGUSR1
(will be used to reload the config) are registered.
Currently we don't get any indication should the daemon crash before
logging is setup.

Signed-off-by: Jan <jan@fajerski.name>