http_sensor: add the initial code for the worker function
Nothing fancy here, Unpack some context and the work items and scrape
the url and send it to kafka, if we have a kafka producer.
This catches all exceptions in order to log them.
Returns a boolean signaling if the work item should be rescheduled or
not. We don't want to reschedule items with an illegal URL but otherwise
a number of potential errors warrant rescheduling.

Signed-off-by: Jan <jan@fajerski.name>